requirements/main.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements/main.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   jupyter-server
    #   openai
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
    #   twisted
autogen-agentchat==0.2.36
    # via waldiez
automat==24.8.1
    # via twisted
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.1.0
    # via nbconvert
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.3.2
    # via requests
constantly==23.10.4
    # via twisted
defusedxml==0.7.1
    # via nbconvert
diskcache==5.6.3
    # via autogen-agentchat
distro==1.9.0
    # via openai
docker==7.1.0
    # via autogen-agentchat
exceptiongroup==1.2.2
    # via anyio
fastjsonschema==2.20.0
    # via nbformat
flaml==2.2.0
    # via autogen-agentchat
fqdn==1.5.1
    # via jsonschema
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via openai
hyperlink==21.0.0
    # via twisted
idna==3.8
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   jsonschema
    #   requests
incremental==24.7.2
    # via twisted
isoduration==20.11.0
    # via jsonschema
jinja2==3.1.4
    # via
    #   jupyter-server
    #   nbconvert
jiter==0.5.0
    # via openai
jsonpointer==3.0.0
    # via jsonschema
jsonschema==4.23.0
    # via
    #   jupyter-events
    #   nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter-client==8.6.2
    # via
    #   jupyter-server
    #   nbclient
jupyter-core==5.7.2
    # via
    #   jupyter-client
    #   jupyter-server
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.10.0
    # via jupyter-server
jupyter-server==2.14.2
    # via waldiez-jupyter (pyproject.toml)
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab-pygments==0.3.0
    # via nbconvert
jupytext==1.16.4
    # via waldiez
markdown-it-py==3.0.0
    # via
    #   jupytext
    #   mdit-py-plugins
markupsafe==2.1.5
    # via
    #   jinja2
    #   nbconvert
mdit-py-plugins==0.4.1
    # via jupytext
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.2
    # via nbconvert
nbclient==0.10.0
    # via nbconvert
nbconvert==7.16.4
    # via jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   jupytext
    #   nbclient
    #   nbconvert
numpy==1.26.4
    # via
    #   autogen-agentchat
    #   flaml
openai==1.44.0
    # via autogen-agentchat
overrides==7.7.0
    # via jupyter-server
packaging==24.1
    # via
    #   autogen-agentchat
    #   jupyter-server
    #   jupytext
    #   nbconvert
pandocfilters==1.5.1
    # via nbconvert
pathvalidate==3.2.1
    # via waldiez-jupyter (pyproject.toml)
platformdirs==4.3.1
    # via jupyter-core
prometheus-client==0.20.0
    # via jupyter-server
ptyprocess==0.7.0
    # via terminado
pycparser==2.22
    # via cffi
pydantic==2.9.0
    # via
    #   autogen-agentchat
    #   openai
pydantic-core==2.23.2
    # via pydantic
pygments==2.18.0
    # via nbconvert
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
python-dotenv==1.0.1
    # via autogen-agentchat
python-json-logger==2.0.7
    # via jupyter-events
pyyaml==6.0.2
    # via
    #   jupyter-events
    #   jupytext
pyzmq==26.2.0
    # via
    #   jupyter-client
    #   jupyter-server
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   docker
    #   tiktoken
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
send2trash==1.8.3
    # via jupyter-server
setuptools==74.1.2
    # via
    #   incremental
    #   zope-interface
six==1.16.0
    # via
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.6
    # via beautifulsoup4
termcolor==2.4.0
    # via autogen-agentchat
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tiktoken==0.7.0
    # via autogen-agentchat
tinycss2==1.3.0
    # via nbconvert
tomli==2.0.1
    # via
    #   incremental
    #   jupytext
tornado==6.4.1
    # via
    #   jupyter-client
    #   jupyter-server
    #   terminado
tqdm==4.66.5
    # via openai
traitlets==5.14.3
    # via
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   nbclient
    #   nbconvert
    #   nbformat
twisted==24.7.0
    # via waldiez
types-python-dateutil==2.9.0.20240906
    # via arrow
typing-extensions==4.12.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   twisted
tzdata==2024.1
    # via pydantic
uri-template==1.3.0
    # via jsonschema
urllib3==2.2.2
    # via
    #   docker
    #   requests
waldiez @ git+https://github.com/waldiez/py.git@17e2b40c86ce6f5bbdc268b6bc4a74f0d33fd124
    # via waldiez-jupyter (pyproject.toml)
webcolors==24.8.0
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.8.0
    # via jupyter-server
zope-interface==7.0.3
    # via twisted
